chore: add build checksum & check in release script
using a simple gulpfile

NOTE:

This update is configured to use SHA-512 which is part of the SHA-2
family ref:

- https://en.wikipedia.org/wiki/SHA-2

This is expected to be good enough for now.

It would have been ideal to use a stronger Keccak-family
algorithm such as keccak 384 or 512 ref:

- https://en.wikipedia.org/wiki/SHA-3#Weakening_controversy

This seems to be part of the OpenSSL code base ref:

- https://github.com/openssl/openssl/pull/16594

but not yet released and not yet supported by Node.js.

Alternative could be to adapt the checksum plugin to use
a pure-JS solution such as:

- https://github.com/phusion/node-sha3
- https://github.com/emn178/js-sha3

PR ref: https://github.com/react-recompose/react-recompose/pull/46